New BarChart and RadarChart
Added a BarChart intended for comparing the average salaries of the jobs. Added a RadarChart for demonstrating the distribution of people in the job market. Refined the PieChart code with multiple colors for better visualization and user experience. Created a new sample dataset for the first graph, the LineChart, to illustrate the salary over time for a particular job. The example job title is AI Specialist.